refactor InvSearchModule and Utils to centralize item search logic and improve inventory handling
